% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{accuracy}
\alias{accuracy}
\title{Classification accuracy}
\usage{
accuracy(pred, truth)
}
\arguments{
\item{pred}{a binary or logical vector of predicted classes}

\item{truth}{a binary or logical vector of observed classes}
}
\value{
accuracy
}
\description{
accuracy(pred, truth) takes inputs of logical or binary vectors of predicted and true classes
and returns the accuracy of the predictions, calculated as
the sum of correct positive classifications and correct negative classifications
divided by the total number of observations
}
\examples{
accuracy(c(1, 1, 1, 0, 0, 0), c(1, 0, 1, 0, 0, 0))
accuracy(c(T, T, T, F, F, F), c(T, F, T, F, F, F))
}
