% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjR2.R
\name{adjR2}
\alias{adjR2}
\title{Adjusted R-squared}
\usage{
adjR2(pred, truth, d)
}
\arguments{
\item{pred}{a binary or logical vector of predicted classes}

\item{truth}{a binary or logical vector of observed classes}

\item{d}{the number of predictors used in prediction}
}
\value{
adjR2
}
\description{
adjR2(pred, truth, d) takes inputs of logical or binary vectors of predicted and true classes as well as
the number of predictors, and returns the adjusted R-squared value
}
\examples{
pred = c(1, 1, 1, 0, 0, 0)
truth = c(1, 0, 1, 0, 0, 0)
adjR2(pred, truth, 1)
}
