% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhAdjust.R
\name{bhAdjust}
\alias{bhAdjust}
\title{Bonferroni-Holm Adjustment}
\usage{
bhAdjust(p, alpha)
}
\arguments{
\item{p}{a vector of p-values}

\item{alpha}{a desired significance level, conventionally .05}
}
\value{
sig.p, a logical vector indicating p-values met significance criteria
}
\description{
bhAdjust(p, alpha) takes a vector of p-values p and a significance level alpha and returns a logical vector
indicating which p-values meet the significance level under the Bonferroni-Holm adjustment criteria
}
\examples{
p = c(0.0001, .00025, .0005, .00075, .001, .005)
alpha = .05
bhAdjust(p, alpha)
}
