% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxMuller.R
\name{boxMuller}
\alias{boxMuller}
\title{Box-Muller transform}
\usage{
boxMuller(n)
}
\arguments{
\item{n}{number of desired pairs}
}
\value{
data frame with columns X1 and X2
}
\description{
boxMuller(n) returns a data frame with columns x1 and x2, representing n pairs of independent, normally-distributed random numbers
}
\examples{
boxMuller(500)
}
