% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chiSquareCounts.R
\name{chiSquareCounts}
\alias{chiSquareCounts}
\title{Chi-Square Counts}
\usage{
chiSquareCounts(tib)
}
\arguments{
\item{tib}{a 2x2 tibble, data frame, or matrix of counts}
}
\value{
chi-square statistic
}
\description{
chiSquareCounts(tib) returns a Chi-squared statistic for the 2x2 contingency table 'tib' by computing
the error between the observed and expected counts
}
\examples{
mydata = data.frame(row.names = c("Success", "Failure"), Group1 = c(21, 3), Group2 = c(14, 10))
chisq = chiSquareCounts(mydata)
}
