% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1.R
\name{f1}
\alias{f1}
\title{F1 score}
\usage{
f1(pred, truth)
}
\arguments{
\item{pred}{a binary or logical vector of predicted classes}

\item{truth}{a binary or logical vector of observed classes}
}
\value{
f1 score
}
\description{
f1(pred, truth) returns the F1 score of the predicted class variable 'pred' and observed classes 'truth'.
'pred' and 'truth' can be either binary or logical vectors. Relies upon ppv() and sensitivity() functions.
}
\examples{
pred = c(1, 1, 1, 0, 0, 0)
truth = c(1, 0, 1, 0, 0, 0)
f1(pred, truth)

pred = c(T, T, T, F, F, F)
truth = c(T, F, T, F, F, F)
f1(pred, truth)
}
