% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimumN.R
\name{minimumN}
\alias{minimumN}
\title{Minimum N}
\usage{
minimumN(d)
}
\arguments{
\item{d}{desired power}
}
\value{
minimum number of observations
}
\description{
minimumN(d) maps a desired power 'd' to a minimum number of observations needed to achieve the desired  power based on a t-test
with assumed Cohen's d of 1
}
\examples{
minimumN(.8)
}
