% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcLollipop.R
\name{pcLollipop}
\alias{pcLollipop}
\title{pcLollipop}
\usage{
pcLollipop(x)
}
\arguments{
\item{x}{a data frame, tibble, or matrix}
}
\value{
plot grid
}
\description{
pcLollipop(x) returns a "lollipop" plot of the loadings of the principal components of 'x'.
'x' can be a matrix, data frame, or tibble with columns corresponding to variables and rows to observations.
For example, if 'x' is a 10x4 matrix, pcLollipop(x) will return a 4x1 grid plot of 10 loadings associated with the 4 principal components.
}
\examples{
data(mtcars)
lolliplot = pcLollipop(mtcars)
}
