% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postHocPower.R
\name{postHocPower}
\alias{postHocPower}
\title{postHocPower}
\usage{
postHocPower(d, n1, n2)
}
\arguments{
\item{d}{desired Cohen's d}

\item{n1}{number of observations in group 1}

\item{n2}{number of observations in group 2}
}
\value{
power estimate
}
\description{
postHocPower(d, n1, n2) returns an estimate of the post-hoc power of 1000 simulated trials with 2 groups with observations n1 and n2 and a Cohen's d of 'd'.
}
\examples{
n1 = 100
n2 = 150
d = 3
php = postHocPower(d, n1, n2)
}
