% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2.R
\name{r2}
\alias{r2}
\title{R-squared}
\usage{
r2(pred, truth)
}
\arguments{
\item{pred}{a binary or logical vector of predicted classes}

\item{truth}{a binary or logical vector of observed classes}
}
\value{
R2
}
\description{
r2(pred, truth, d) takes inputs of logical or binary vectors of predicted and true classes
and returns the R-squared value according to residual and total sum of squares error.
}
\examples{
pred = c(1, 1, 1, 0, 0, 0)
truth = c(1, 0, 1, 0, 0, 0)
r2(pred, truth, 1)
}
