% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{sensitivity}
\alias{sensitivity}
\title{Sensitivity}
\usage{
sensitivity(pred, truth)
}
\arguments{
\item{pred}{a binary or logical vector of predicted classes}

\item{truth}{a binary or logical vector of observed classes}
}
\value{
sensitivity
}
\description{
sensitivity(pred, truth) returns the sensitivity of the predictor according to predicted class 'pred' and observed classes 'truth'.
'pred' and 'truth' can be either binary or logical vectors.
Sensitivity is computed as the number of correct positive predictions divided by the total number of positive predictions.
}
\examples{
pred = c(1, 1, 1, 0, 0, 0)
truth = c(1, 0, 1, 0, 0, 0)
sensitivity(pred, truth)

pred = c(T, T, T, F, F, F)
truth = c(T, F, T, F, F, F)
sensitivity(pred, truth)
}
