% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificity.R
\name{specificity}
\alias{specificity}
\title{Specificity}
\usage{
specificity(pred, truth)
}
\arguments{
\item{pred}{a binary or logical vector of predicted classes}

\item{truth}{a binary or logical vector of observed classes}
}
\value{
specificity
}
\description{
specificity(pred, truth) returns the specificity of the predictor according to predicted class 'pred' and observed classes 'truth'.
'pred' and 'truth' can be either binary or logical vectors.
Specificity is computed as the number of correct negative predictions divided by the total number of negative predictions.
}
\examples{
pred = c(1, 1, 1, 0, 0, 0)
truth = c(1, 0, 1, 0, 0, 0)
specificity(pred, truth)

pred = c(T, T, T, F, F, F)
truth = c(T, F, T, F, F, F)
specificity(pred, truth)
}
