% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoSidedZ.R
\name{twoSidedZ}
\alias{twoSidedZ}
\title{Two-Sided Z}
\usage{
twoSidedZ(z)
}
\arguments{
\item{t}{a test-statistic}
}
\value{
area
}
\description{
twoSidedZ(z) takes a test-statistic 'z' and returns the area under the normal distribution >= |z|
}
\examples{
z = 4
area = twoSidedZ(z)
}
