% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unscale.R
\name{unscale}
\alias{unscale}
\title{Unscale}
\usage{
unscale(mat)
}
\arguments{
\item{mat}{a scaled object}
}
\value{
unscaled data
}
\description{
unscale(mat) takes a scaled object input and rescales and recenters data to its original mean and standard deviation.
}
\examples{
mat = mtcars
mat.scaled = scale(mat)
mat.unscaled = unscale(mat.scaled)

}
