% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/welchT.R
\name{welchT}
\alias{welchT}
\title{Welch's T}
\usage{
welchT(x, y)
}
\arguments{
\item{x}{a numerical vector}

\item{y}{a numerical vector}
}
\value{
Welch's t-statistic
}
\description{
welchT(x, y) takes numerical vectors 'x' and 'y' and returns a Welch's t-statistic between the two groups.
}
\examples{
x = rnorm(n = 50, mean = 0, sd = 1)
y = rnorm(n = 50, mean = 0, sd = 2.5)
t.stat = welchT(x, y)
}
